input in ms > change int to long

+++ src/main/java/com/sparking/service/FieldService.java
@@ -21,7 +21,7 @@ public interface FieldService {
 
     boolean managerDelete(int id, String phone);
 
-    List<FieldAnalysis>analysis(int fieldId, int since, int until, String unit) throws ParseException;
+    List<FieldAnalysis>analysis(int fieldId, long since, long until, String unit) throws ParseException;
 
     FieldJson data2Json(Field field);
